+++ hutool-core/src/test/java/cn/hutool/core/convert/NumberChineseFormaterTest.java
@@ -52,7 +52,11 @@ public class NumberChineseFormaterTest {
 	public void digitToChineseTest() {
 		String digitToChinese = Convert.digitToChinese(12412412412421.12);
 		Assert.assertEquals("壹拾贰万肆仟壹佰贰拾肆亿壹仟贰佰肆拾壹万贰仟肆佰贰拾壹元壹角贰分", digitToChinese);
+		
 		String digitToChinese2 = Convert.digitToChinese(12412412412421D);
 		Assert.assertEquals("壹拾贰万肆仟壹佰贰拾肆亿壹仟贰佰肆拾壹万贰仟肆佰贰拾壹整", digitToChinese2);
+		
+		String digitToChinese3 = Convert.digitToChinese(2421.02);
+		Assert.assertEquals("贰仟肆佰贰拾壹元零贰分", digitToChinese3);
 	}
 }